DjangoServer/requirements.txt
aniso8601==8.0.0
beautifulsoup4
certifi==2020.4.5
chardet==3.0.2
click==7
dlib
django-cors-headers
djangorestframework
fastapi
folium==0.12.1
Flask-RESTful==0.3.8
Flask==1.1.2
googlemaps==4.4.5
html5lib
icecream==2.1.1
idna==2.10
imblearn
isort==5.9.3
itsdangerous==1.1.0
JPype1
jsonify==0.5
jupyter==1
keras
keras-preprocessing
konlpy==0.5.2
lxml
MarkupSafe==2.1.1
matplotlib
mccabe
mysql-connector
mysqlclient
nltk
numpy
opencv-python
pandas
pandas-datareader==0.9.0
pep8==1.7.1
plotly==5
prophet==1.1.1
python-dateutil==2.8.1
pytz==2020.5
requests==2.24.0
seaborn
selenium==3.141.0
six==1.15.0
scikit-learn
SQLAlchemy==1.4.39
tensorflow-gpu==2.9.3
tensorflow==2.9.1
tensorflow_datasets
torch
torchvision
torchaudio --extra-index-url https://download.pytorch.org/whl/cu117
tweepy==3.10.0
tqdm
urllib3==1.25.10
uvicorn
webdriver-manager
Werkzeug==2.2
wincertstore==0.2
wordcloud
yfinance==0.2.3
PyAutoGUI
